+++ src/components/boilerplats/Header.jsx
@@ -1,22 +1,22 @@
 import "@fontsource/pacifico"
 import "@fontsource/pacifico/400.css"
 import "@fontsource/pacifico/cyrillic-400.css"
-import '../stylesheet/header.css'; 
+import '../stylesheet/Header.css';
 
 const Header = () => {
-  return (
-    <header>
-    <div className="logo">Bidmaster</div>
-    <nav>
-      <ul className="nav-links">
-        <li><a href="#">Lägg till annons</a></li>
-        <li><a href="#">Skapa konto</a></li>
-        <li><a href="#">Logga in</a></li>
-        <li><a href="#">English</a></li>
-      </ul>
-    </nav>
-  </header>
-  );
+    return (
+        <header>
+            <div className="logo">Bidmaster</div>
+            <nav>
+                <ul className="nav-links">
+                    <li><a href="#">Lägg till annons</a></li>
+                    <li><a href="#">Skapa konto</a></li>
+                    <li><a href="#">Logga in</a></li>
+                    <li><a href="#">English</a></li>
+                </ul>
+            </nav>
+        </header>
+    );
 };
